+++ guillotina/tests/test_api.py
@@ -583,10 +583,21 @@ async def test_items(container_requester):
         response, status = await requester('GET', '/db/guillotina/@items?page_size=10')
         assert len(response['items']) == 10
         assert response['total'] == 22
+        items = [i['UID'] for i in response['items']]
 
         response, status = await requester('GET', '/db/guillotina/@items?page_size=10&page=2')
         assert len(response['items']) == 10
         assert response['total'] == 22
+        items.extend([i['UID'] for i in response['items']])
 
         response, status = await requester('GET', '/db/guillotina/@items?page_size=10&page=3')
         assert len(response['items']) == 2
+        items.extend([i['UID'] for i in response['items']])
+
+        # we should have 22 unique uids now
+        assert len(set(items)) == 22
+
+        response, status = await requester(
+            'GET', '/db/guillotina/@items?omit=guillotina.behaviors.dublincore.IDublinCore')
+        item = response['items'][0]
+        assert 'guillotina.behaviors.dublincore.IDublinCore' not in item